connected home page
also reorganized folders
