--- marketplace-tutorial-master/README.md
@@ -1,26 +1,0 @@
-# marketplace-tutorial
-Python / Flask Workshop Series Code
-
-This marketplace application is using the flask web framework and was used to teach an intro-level python workshop
-
-### To run this application locally:
-
-1. Clone the repo
-2. Make sure you have Python 3.6 and flask installed
-
-Download Python 3.6: https://www.python.org/downloads/release/python-360/
-
-```
-pip install -U Flask
-
-```
-
-3. Make sure you are in the /marketplace-tutorial directory and run the following commands in your terminal:
-
-```
-export FLASK_APP=marketplace
-export FLASK_ENV=development
-flask run
-```
-
-4. Navigate to the url provided in the terminal to see the web application
