--- marketplace-tutorial-master/marketplace/templates/store/index.html
@@ -1,31 +1,0 @@
-{% extends 'base.html' %}
-
-{% block header %}
-
-{% endblock %}
-
-{% block content %}
-<div class="container">
-  <div class="row">
-    {% for item in items %}
-      <div class="col-4 card">
-        <img height=auto width=300 class="card-img-top" src="../../static/img/{{ item['item_image'] }}" alt="Card image cap">
-        <div class="card-body">
-          <h5 class="card-title">{{ item['item_name'] }}</h5>
-          <p class="card-text">{{ item['item_description']}}</p>
-          <div>
-            <form action="{{ url_for('cart.add_cart', item_id=item['id']) }}" method="post">
-              <input class="btn btn-primary" type="submit" value="Add to Cart">
-            </form>
-            {% if g.user['admin']%}
-              <form action="{{ url_for('store.delete', item_id=item['id']) }}" method="post">
-                <input class="btn btn-danger" type="submit" value="Remove">
-              </form>
-            {% endif %}
-        </div>
-        </div>
-      </div>
-    {% endfor %}
-  </div>
-</div>
-{% endblock %}
